% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCCS.R
\name{QCCS}
\alias{QCCS}
\title{Quality Control Columns Set}
\value{
Object of class QCCS
}
\description{
The p value is calculated using the multinomial distribution from vectors and their corresponding probabilities.
}
\examples{

## ------------------------------------------------
## Method `QCCS$new`
## ------------------------------------------------

vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob)


## ------------------------------------------------
## Method `QCCS$QCCS`
## ------------------------------------------------

vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob,Source="Ariza et al.,2019")
A$QCCS()

}
\references{
\insertRef{QCCS}{PaolaR6Nuevo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QCCS-new}{\code{QCCS$new()}}
\item \href{#method-QCCS-QCCS}{\code{QCCS$QCCS()}}
\item \href{#method-QCCS-clone}{\code{QCCS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-new"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$new(vectors, prob, ID = NULL, Date = NULL, Source = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vectors}}{vector list.}

\item{\code{prob}}{probabilities list.}

\item{\code{ID}}{Identifier. By default ID is a date in YYYYMMDD format}

\item{\code{Date}}{Date provided by the user. By default the date provided by the system will be taken.}

\item{\code{Source}}{Indicates where the matrix comes from (article, project, etc.). By default is NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-QCCS"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-QCCS}{}}}
\subsection{Method \code{QCCS()}}{
Public method that, using a list of vectors and their corresponding probabilities, through a multinomial distribution, calculates the p value using each of the vectors. The reference \insertCite{QCCS}{PaolaR6Nuevo} is followed for the computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$QCCS()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The p value is obtained for each vector, and using the Bonferroni criterion it is decided whether the elements are well classified or not.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob,Source="Ariza et al.,2019")
A$QCCS()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
