% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCCS.R
\name{QCCS}
\alias{QCCS}
\title{Quality Control Columns Set}
\value{
Object of class QCCS
}
\description{
The p value is calculated using the multinomial distribution from vectors and their corresponding probabilities.
}
\examples{
vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob)

## ------------------------------------------------
## Method `QCCS$QCCS`
## ------------------------------------------------

vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob)
A$QCCS()
}
\references{
Ariza-López, F. J., Rodríguez-Avi, J., Alba-Fernández, M. V., & García-Balboa, J. L. (2019). Thematic accuracy quality control by means of a set of multinomials. Applied Sciences, 9(20), 4240.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QCCS-new}{\code{QCCS$new()}}
\item \href{#method-QCCS-QCCS}{\code{QCCS$QCCS()}}
\item \href{#method-QCCS-clone}{\code{QCCS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-new"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$new(vectors, prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vectors}}{vector list}

\item{\code{prob}}{probabilities list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-QCCS"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-QCCS}{}}}
\subsection{Method \code{QCCS()}}{
Using a list of vectors and their corresponding probabilities, through a multinomial distribution, the p value is calculated using each of the vectors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$QCCS()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The p value is obtained for each vector, and using the Bonferroni criterion it is decided whether the elements are well classified or not.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vectors<-list(c(47,4,0),c(40,5,3),c(45,6,2),c(48,0))
prob<-list(c(0.95,0.04,0.01),c(0.88,0.1,0.02),c(0.9,0.08,0.02),c(0.99,0.01))
A <- QCCS$new(vectors,prob)
A$QCCS()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCCS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QCCS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCCS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
