% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\value{
Object of class test
}
\description{
Creates a new instance of this [R6][R6::R6Class] class.
}
\examples{

## ------------------------------------------------
## Method `test$new`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")


## ------------------------------------------------
## Method `test$StHell`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$StHell()


## ------------------------------------------------
## Method `test$kappa.test`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$kappa.test()


## ------------------------------------------------
## Method `test$OverallAcc.test`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$OverallAcc.test()


## ------------------------------------------------
## Method `test$Tau.test`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$Tau.test()


## ------------------------------------------------
## Method `test$TSCM.test`
## ------------------------------------------------

A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Garcia-Balboa el al. 2018")
f$TSCM.test()

}
\references{
\insertRef{garcia2018}{PaolaR6Nuevo}

\insertRef{ma1995Tau}{PaolaR6Nuevo}

\insertRef{alba2020}{PaolaR6Nuevo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-test-new}{\code{test$new()}}
\item \href{#method-test-StHell}{\code{test$StHell()}}
\item \href{#method-test-kappa.test}{\code{test$kappa.test()}}
\item \href{#method-test-OverallAcc.test}{\code{test$OverallAcc.test()}}
\item \href{#method-test-Tau.test}{\code{test$Tau.test()}}
\item \href{#method-test-TSCM.test}{\code{test$TSCM.test()}}
\item \href{#method-test-clone}{\code{test$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-new"></a>}}
\if{latex}{\out{\hypertarget{method-test-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$new(A = NULL, B = NULL, ID = NULL, Date = NULL, Source = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{A}}{Matrix}

\item{\code{B}}{Matrix}

\item{\code{ID}}{Identifier. By default ID is a date in YYYYMMDD format}

\item{\code{Date}}{Date provided by the user. By default the date provided by the system will be taken.}

\item{\code{Source}}{Indicates where the matrix comes from (article, project, etc.). By default is NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-StHell"></a>}}
\if{latex}{\out{\hypertarget{method-test-StHell}{}}}
\subsection{Method \code{StHell()}}{
Public method that provides the Hellinger distance. The reference \insertCite{garcia2018}{PaolaR6Nuevo} is followed for the computations.
The mathematical expression is:

\deqn{
HD = \dfrac{4nm}{n+m} \sum^{M}_{i=1} (\sqrt{p_i}-\sqrt{q_i})^2
}

Where:
\enumerate{
  \item HD: Hellinger Distance
  \item n: number of elements in the matrix A.
  \item \code{p_i}: element i of the probability vector of matrix A.
  \item \code{q_i}: element i of the probability vector of matrix B.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$StHell(A = NULL, B = NULL, p = NULL, q = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{A}}{matrix. By default, the defined matrix is taken to create the object of the test class.}

\item{\code{B}}{matrix. By default, the defined matrix is taken to create the object of the test class.}

\item{\code{p}}{matrix probability vector. By default, the probability of success for each cell is taken.}

\item{\code{q}}{matrix probability vector. By default, the probability of success for each cell is taken.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The statistic value of the statistical test based on the Hellinger distance.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$StHell()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-kappa.test"></a>}}
\if{latex}{\out{\hypertarget{method-test-kappa.test}{}}}
\subsection{Method \code{kappa.test()}}{
Public method that tests whether two independent confusion matrices are significantly different using their kappa index. For the calculations, the reference \insertCite{congalton2008}{PaolaR6Nuevo} is followed.
The mathematical expression to calculate its statistic is:

\deqn{
Z = \dfrac{|K1-K2|}{\sqrt(var(K1)+var(K2))}
}

Where:
\enumerate{
  \item K1: kappa index of matrix A
  \item K2: kappa index of matrix B
  \item var(K1): variance of K1.
  \item var(K2): variance of K2.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$kappa.test(alpha = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{significance level. By default alpha=0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the value of the statistic between kappa values and its z score for a given alpha significance level.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$kappa.test()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-OverallAcc.test"></a>}}
\if{latex}{\out{\hypertarget{method-test-OverallAcc.test}{}}}
\subsection{Method \code{OverallAcc.test()}}{
Public method that tests whether two independent confusion matrices are significantly different using their overall accuracy index. For the calculations, the references \insertCite{book,ma1995Tau}{PaolaR6Nuevo} is followed.
The mathematical expression to calculate its statistic is:

\deqn{
Z = \dfrac{k1-k2}{\sqrt{var(k1)+var(k2)}}
}

Where:
\enumerate{
  \item k1: overall index of matrix A
  \item k2: overall index of matrix B
  \item var(K1): variance of K1.
  \item var(K2): variance of K2.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$OverallAcc.test(alpha = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{significance level. By default alpha=0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of the statistic's value between the overall accuracies and its z-score for a given alpha significance level.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$OverallAcc.test()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-Tau.test"></a>}}
\if{latex}{\out{\hypertarget{method-test-Tau.test}{}}}
\subsection{Method \code{Tau.test()}}{
Public method that tests whether two independent confusion matrices are significantly different using their Tau index. For the calculations, the references \insertCite{book,ma1995Tau}{PaolaR6Nuevo} is followed.
The mathematical expression to calculate its statistic is:

\deqn{
Z = \dfrac{k1-k2}{\sqrt{var(k1)+var(k2)}}
}

Where:
\enumerate{
  \item k1: Tau index of matrix A
  \item k2: Tau index of matrix B
  \item var(K1): variance of K1.
  \item var(K2): variance of K2.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$Tau.test(alpha = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{significance level. By default alpha=0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of the statistic's value between the Tau index and its z-score for a given alpha significance level.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Congalton and Green 2008")
f$Tau.test()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-TSCM.test"></a>}}
\if{latex}{\out{\hypertarget{method-test-TSCM.test}{}}}
\subsection{Method \code{TSCM.test()}}{
Public method that performs a homogeneity test between two matrices based on the Hellinger distance. The reference \insertCite{garcia2018}{PaolaR6Nuevo} is followed for the computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$TSCM.test(n1 = NULL, alpha = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n1}}{Number of bootstraps that you want to generate. By default n=10000.}

\item{\code{alpha}}{significance level. By default alpha=0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
p value and decision to make.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),nrow=4,ncol=4)
B<-matrix(c(45,6,0,4,4,91,8,7,12,5,55,3,24,8,9,55),nrow=4,ncol=4)
f <- test$new(A,B,Source="Garcia-Balboa el al. 2018")
f$TSCM.test()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test-clone"></a>}}
\if{latex}{\out{\hypertarget{method-test-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{test$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
